tests/submission/PhilipTaylor/canvas/2d.line.cap.invalid.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.line.cap.invalid</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.line.html">line</a>.<a href="index.2d.line.cap.html">cap</a>.invalid</h1>
<p class="desc">Setting lineCap to invalid values is ignored</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.lineCap.invalid">2d.lineCap.invalid</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Setting lineCap to invalid values is ignored");
_addTest(function(canvas, ctx) {

ctx.lineCap = 'butt'
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'invalid';
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'ROUND';
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'round\0';
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'round ';
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = "";
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");

ctx.lineCap = 'butt';
ctx.lineCap = 'bevel';
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");


});
</script>

<div id="log"></div>
</body></html>
